refactor(CarMoveResult, GameInfo): 패키지 변경
