+++ src/test/java/racingcar/domain/car/ParticipateCarsTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.car.strategy.FixedMovementStrategy;
+import racingcar.vo.car.CarMoveResult;
 
 import java.util.ArrayList;
 import java.util.Arrays;
