+++ src/test/java/racingcar/domain/game/GameResultTest.java
@@ -3,15 +3,12 @@ package racingcar.domain.game;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.car.CarMoveResult;
+import racingcar.vo.car.CarMoveResult;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 class GameResultTest {
