+++ src/test/java/racingcar/domain/game/PhaseResultTest.java
@@ -3,10 +3,9 @@ package racingcar.domain.game;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.CarMoveResult;
+import racingcar.vo.car.CarMoveResult;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
