+++ src/test/java/racingcar/vo/game/GameInfoTest.java
@@ -1,8 +1,10 @@
-package racingcar.domain.game;
+package racingcar.vo.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.vo.game.GameInfo;
+
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 class GameInfoTest {
